Add svg module
About to be rewritten anyway, but here's an initial approach to use JS
functions. This doesn't work so well, tho, as the SVG needs to be
rendered. A work-around to not have it visible is to position it
absolutely at 0, 0 and give it an opacity of 0, but even then the
document needs to be rendered before the values can be used.