Improve anti-aliasing issue slightly
For curved lines, introduced by line styles, the overlap in its present
form won't work, because a line orthogonal to the line being drawn with
the style might intersect the fancy line in more than one point, which
would utterly break any rendering that just translates the line along
that direction.
Hence then anti-aliasing issue is reintroduced. Rendering the field's
shape with a stroke of the same colour as the tincture improves the
issue a little.