Add outline mode
This won't be the complete solution, since it requires the distinction
between ordinary and division lines, because the outlines are rendered
inside the field's mask, showing only half the path. For division lines
that's fine, since the other half will be rendered by the adjacent
field, but ordinaries only get one half, making them look half as thick.

The CSS addresses this, but it also results in the path then being
thicker inward, which in particular shows when an edge is part of the
edge of the parent field.

A better way will be to explicitly render the relevant paths over the
masked field, then they also all can have a uniform thickness, and paths
outside of the visible area don't need to be constructed.