Improve speed by reducing bounding box accuracy
Always calculating a thousand points is too slow once there are several
fields on the screen, this is almost as accurate and much, much faster.
In the future there hopefully is an even better way to do this.

Many sub-fields even know their bounding box, they wouldn't need this
calculation at all.