Fix copy button, semi-fix textarea error
The copy command was accidentally lost in a recent editing frenzy.
The textarea complained about having a value but no on-change event.
This handler tries to write the value back into it, but it doesn't seem
to work properly. However, with copy/paste buttons there this isn't so
pressing right now.