Fix field re-use when switching from/to tierced divisions
A tierced division had the third field set with a tincture, switching to
a division with more than 2 fields but only 2 mandatory fields would
keep that third tincture for the third field, which makes little sense.

Reversely, coming from such a division (e.g. quarterly) has the third
field set up as a reference of the first, which would then be kept when
switching to a tierced division. However, tierced divisions don't expect
references, since they are all mandatory, so this breaks.

To explain why this re-using of fields is done to begin with: it is a
much nicer user experience to be able to change division after already
changing its fields. Otherwise any division would result in the default
fields, making changes really tedious.

The easier editing comes at the price of potentially introducing this
sort of extra complexity and potential for unexpected behaviour.