Better separate the concepts of field and field environment
The field will be any blazonry-relevant description, while the field
environment is merely the space the dimensions, the meta data into which
the field is rendered.

Previously "field" was used for the environment and here and there for
the actual field as well, which is confusing and makes some semantics
dodgy.

This also paves the way for renaming the "content" keywords inside
fields and ordinaries again, which didn't fit so well.

Finally, it'll be useful for the proper separation of meta data that
must be included in the field, such as how the fess point should be
calculated, and that which belongs in the environment, such as where the
fess point actually is, as that would not be blazoned.