Add "none" tincture to have a better default
Especially ordinaries that way won't be invisible if the fields below
happen to have the same colours as their default. This still might
happen if the fields also have no tincture set, but in that case I think
a best effort has been made. A solution for this also might be selective
outlining, but that's a future feature.

For divisions the second/third field still gets a tincture to
distinguish it from the unfilled field, but that first field should be
enough to indicate a change, even if those default tinctures are found
on the fields below or around.