Pass bounding box information down to sub environment
This will make positioning and division calculations for the sub fields
more accurate.

However, it immediately introduces a box, since it now doesn't take the
bounding box from the actual path anymore, but rather an ideal path.
This works for all straight lines, but for line styles that extend into
the negative on the y axis it introduces gaps, since the mask and field
now don't reach all those spaces formed by the line anymore.

The fix for that will add quite a bit of complexity, but it'll be worth
it to have more control over the ideal boundaries.