Remove starting point from infinity path
The starting point usually is the end of a line already. Due to the
line's extra length, this point usually will be "past" the point the
line was aiming for. But that would be the point this sets as the first,
resulting in potentially ugly self-intersections, and jumping back into
an area that might still be rendered due to the parent's line style.

Removing the point should be safe. If it is really needed, then the
calling context can set it specifically.