Remove top-level-render from charge/render
I don't foresee needing that for charges, since they don't include a new
field that might require rendering, unlike ordinaries.

Mobile ordinaries might not fit in the current ordinary mould, since
they need positioning attributes, but even if that is the case, then
they should be treated separately, and not as charges, to keep the code
simple.