Allow removing the outline of charges
This should be a nice option, depending on the charge and the style of
the coat of arms overall. Most charges probably look better with
outline, but simple ones like the roundel likely will look better
without it in many cases.

Perhaps stroke-width also can be adjusted at some point, so it is
consistent with the environment, even if the charge is scaled.

Finally, if the global outline mode is enabled, then that should
override this charge hint.