Speed up SVG to EDN processing
The lein process starts up slowly, quickly making this script very slow,
as the number of SVG files grows. This change calls it once with all SVG
files, which is a good deal faster. However, it means all the results
share the same output directory, so the SVG files must be unique
globally. I think that'll happen anyway with the current naming scheme,
unless a charge name exists in different groups.

If that happens at some point, I'll figure out a remedy.