Rename variants "1" to "default"
Also reduce the variant key/name to not include redundant information
from the charge and attitude. The default variant also drops the
"variant" prefix, so it is only "default". I think that's much easier to
read in the interface, and it also makes for a neater :variant value in
the charge data.