Fix charge mask handling of black-filled paths
These should be treated as outline, but the SVG optimizer removed
defaults such as fill:#000, so they couldn't be excluded in the mask by
setting the fill to "none".

svgo allows disabling that setting, which fixes the issue for the
existing charge SVGs, but it might come back if an SVG doesn't
explicitly state the fill colour. If it shows up again, then I can try
to be smarter about this, but it's not so straight-forward, so I'll
leave it at that for now.