Add clip-path for charge
This allows pointer events to reach underlying fields, at least outside
of the rectangle the charge occupies. Due to a bug several current
browsers don't allow clicks through transparent masks, so selecting the
charge or underlying fields is not as accurate as it should be, but this
at least allows reaching the other fields outside of the rectangle.