Add ermine furs
Vair might need some extra work, it looks like it's not uncommon to
specify the number of vair elements and to expect them to stretch to
fill an ordinary.
This might happen with ermine as well, actually. So there might be a
need to treat them as a form of paly division.

I'll leave it at ermine-based furs for now.