Add more line options and a submenu interface for them
Not all line usages will work with default settings. At times lines will
meet at certain angles in some divisions or ordinaries and might need
tweaking. There also might be different styles left open for
interpretation of the blazon, such as the number of repetitions of the
line pattern and its eccentricity.

This is not done, as the options are not always applicable and need to
be validated. It also doesn't play well with the "reverse" lines used so
far to make lines end in certain points. That will need some new maths.