Refactor line construction, correct offset/flipped
Some divisions and ordinaries didn't use the right line alignment,
offset, and flipping, it's more consistent and logical now.

The maths also were not up to snuff in some instances, but now it should
be pretty logical. The offset helped pin point where the lines were not
the right way around or just accidentally looked correct.