Fix edge cases with line offsets
It's actually wrong to ever use "m" there, since it breaks the path,
even if it ends up on the same point, making it close wrongly. This
breaks the cross ordinary at offset 0.

It's actually fine to have "l 0,0" in there, if there is no offset, so
simplify the whole thing and always add an offset line at the beginning
and end.