Fix selection of nested fields, especially ordinaries
The db-path handed down to an ordinary field was created like that of a
division, which has :fields <idx> to identify each sub field, but
ordinaries currently only have :field.

In the future ordinaries might have multiple fields, especially when
they show up in groups. For now the clumsy check of the type should do.

Finally, the selection indicator was rendered on top of sub components,
which is not ideal.