Add distinction of mobile subordinaries
This also makes it easier to distinguish them for when positioning
groups become available, which wouldn't apply to ordinaries.
In the future there also are going to be fixed subordinaries, which
might be treated differently from ordinaries in some contexts, but at
the very least also shouldn't be considered mobile.