Improve expanding/collapsing of components
Closing a component will recursively close its children, so re-opening
it will show the structure with less confusion.
Similarly, opening a component will open all parent components, so it is
definitely visible. This makes most sense when selecting a component via
the coat of arms, which might not be visible at the time. Now it opens
the entire tree up to the selected component.