Add option to flip a line style vertically
This is particularly useful for :other-line, which is on the opposite
edge, and flipping it results in following the top line at a constant
distance.

It needs to be a select element to allow the third state of "inherit",
which is nil in the data structure. A second checkbox like there is for
sliders isn't a good UI. This also isn't ideal, and it is strange for a
normal line, where there only are on and off, but at least it is
consistent.