Change namespace and reorganize sources
With other modules about to be added, I think it makes sense to separate
them into several shadow-cljs targets with their own dependencies.

That way all the CLJS code can be shared, but the builds don't have to
include all of it.

There still will be some duplication, for instance with the CSS, but
maybe there'll be a nice way to do that as well.