Implement sign-up and confirmation process in modals
Also complete the proper login afterwards, if confirmation was needed.
This now also works if one leaves the site and comes back and tries to
login. In that case it is slightly annoying that one has to provide
credentials after the confirmation again, but I think that's just the
workflow, and the advantage is that the login path is the only one
completing a login.