Store SVG and EDN data in a sha1-named S3 object
That way there's no duplication if the SVG/EDN data didn't change. It'll
also mean clients have it cached, if only the name changes for instance.

But the current implementation of using subscriptions like events is
really ugly. That'll need some work, just wanted to get something
working for now.