Refactor user and charge forms to use a db-path
The previous approach with form-id was not as flexible and very
different from the way it is already done in coat-of-arms, which is
closer to the way this patch is doing it. Each form simply has a
db-path, its fields can be accessed by conjing to that db-path.

Form errors are stored elsewhere, spanning the same map structure based
on the db-path, so they don't change the actual form data.

Finally, modals are given an on-cancel handler, so they can clear the
form or whatever is necessary when they are closed. The auth-related
modals in frontend.user use this.