Move extra tinctures into submenu for charge form
The supported tinctures must come from somewhere else, now that there
won't be a charge-map in the same way as before anymore.
And indeed it makes sense to specify those tinctures, even if the
selected charge doesn't support them. In the future there should be some
warning in that case.

For now just list all supported extra tinctures to unblock the feature,
even if a lion rarely will be beaked.