Add API function to generate a charge-map
This will be similar to the current static charge-map, but the grouping
of attitude, facing, and variants changes. This likely will evolve quite
a bit anyway, and at some point it won't be feasible to do this on the
fly with an API call.

An asynchronous job can generate the data accessible by the public (or
other groups) and put them on S3 for caching and download, but that can
be done when this file is a bit larger.

For now it is convenient that it returns public charges and one's own,
which of course should be accessible.