Allow specifying :length as number of desired points
For the squiggly algorithm it is useful to have a number of points that
depends on the length of the path, so the shape of it will be squiggled
at a somewhat constant pace. If the path becomes too long, then a fixed
number would result in too much degradation.