Fix line stitching
With the new path from svg-path-properties, which is not as exact as the
browser one, it seems, line stitching broke, because the "M" segment in
the beginning was not properly removed. The regex was flawed and also
didn't include 'e' as a character to be expected, but numbers like
"1e-9" showed up instead of 0.

Also, if the second number is negative, then there isn't necessarily a
space or a comma between them.