Make component opening possible for any paths
Not all the paths lead to a map that has a :component set. This was
required back when the :ui data was stored inside the coat-of-arms data,
but now it is separate, so it's fine to also mark paths as "open" that
don't have a component representation.
In particular this allows opening the "attribution" component that was
just added.