Refactor full URL creation, move it all into heraldry.util
That way frontend and backend can use the same functions. It also
doesn't use reitit anymore, so the backend doesn't need the routing
setup to create the URLs.

This is a little annoying, as it means routes and the URLs could
diverge, but at least it's only in one place and frontend and backend
will always agree.