Unify the async data fetching process
There was a lot of code duplication with the :loading and whatnot. Now
it's all done in one place, which also handles caching of multiple
different queries going to the same db-path, and it also notices if the
content belongs to a different query, making it much safer for switching
views that might share some of these db-paths.