Draw saltire and gyronny diagonals past the environment edges
This solves some issues with extreme cases, for instance a saltire in a
bend with the origin very close to the top-left, which has a line style
that reaches above and below the straight line. That would show a
straight line past the saltire reaching the edge of the calculated
environment in the top left corner.