Allow nil as range default and prevent it from using the minimum
Not really used right now, but I thought about using it for chequy when
the vertical number should be based on the horizontal number, resulting
in square parts.
That's tricky, tho, because the default fields won't know at that point
how many parts there will be, so it isn't used for now.