Turn current chequy into a general quarterly
This will be separate from the 2x2 quarterly, as that one supports line
styles and an origin, which makes less sense for higher number quarterly
divisions.

At the same time, allow 2x2 in the general quarterly as well, but
without the origin and field filling logic, as that might be the right
thing for some arms.