Make all non-straight line styles use the new function
This doesn't fully work with negative y offsets for joint angled lines.
Without outline the result looks decent for most non-edge-cases, but the
outline is broken for inner joints.

The straight style also is separate and repeats a lot of code, but it is
the one that'll work properly for all joints for now.

More logic is needed to tidily cut off fimbriated line lines somewhere
in the middle of a pattern and then join them at that point.

Lastly, non-function line styles don't have proper fimbriation in
x-direction yet and look weird.