Solve fimbriation issues with negative/positive offsetting
This was the missing piece to deal with fimbriation collision in inner
joints, where the "lower" lines now get extra offsetting to avoid
shifting of the line pattern.

Ideally the fimbriation lines would properly intersect with the angled
fimbriation lines and make the use of these offset straight lines
unnecessary, but that's more complicated. This provides a stable and
consistent solution to render fimbriation lines in inside/outside/even
mode, for inner and outer joints.