Move form-for-fimbriation into its own, reusable function
This is not perfect, since it doesn't display/handle inheritance from
the opposite-line well. It is impossible to go back to the default, for
instance, and the inherited tinctures are not displayed.
Right now I don't have a better approach, and it's an edge case that can
be addressed later.