Improve line/render, so it can take multiple line segments
Now multiple line-data maps can be stitched together, and the
fimbriation will be handled properly as well.
For this the line up/down masks need to be merged as well, but in a way
that doesn't overextend them for inner joint angles and doesn't leave
gaps for outer joint angles. The solution is to find the intersection
points of the mask lines parallel to the line segments.

Also then use this function for ordinaries with a joint: chevron, cross,
and saltire.