Allow alpha gradients for shadow/highlights
This complects the code quite a bit, but it should be worth it to allow
optional shading/highlighting with gradients.

Known bug: fimbriation will fimbriate a ghost shape of gradient shapes
that extend past the charge border, if they don't have an explicit
stroke color set, because it'll set a default for the stroke colour and
then render that shape, even if fill is already overridden to "none".
This is not trivially fixed, but it is an edge case I can live with
right now, as the shading can be expected to be inside the charge.