Add pile ordinary
The configuration is rather complex, but the origin/anchor system allows
any pile one could want to add. In addition to that it is the first
ordinary that actively determines points on the surrounding environment
shape (or parent shapes), so it can place the point precisely on the
edge.

It also introduces a new position option:
- type: :point to use the anchor point as target point
        :edge to use origin and anchor to project a point on the edge of
              the environment, this is the default for anchors with
              :angle as the point setting, because they don't define an
              anchor point