Fix heights of two escutcheons
Creating two piles from top-left, top, top-right, anchoring in bottom,
reveals that the bottom point was not on chosen well for these two yet.
for the Czech one the exact calculation isn't trivial, but trial and
error resulted in a good approximation that looks very, very close.
The English one has that little extra corner at the bottom, the shift
puts the bottom point into that point of the corner, so a vertical pile
throughout will point all the way down to it.