Add pile ordinary thickness configuration by angle
This is tricky, and not immediately intuitive, but when using the angle
configuration together with an origin alignment of left or right edge,
then the diagonal the pile points along depends on the stretch factor
and the angle, such that a shortened or extended pile always point
through or toward the anchor point AND keep the point angle at the
desired angle AND align the edge to the origin point.

This approach calculates the required values by constructing the
situation  with a "known" length of 1 between calculated-origin and
calculated-point, which results in some edge length, which needs to
match our known edge-length from real-origin to real-anchor, a quotient
of those two then is a factor the whole construction has to be
multiplied with, such that the edge-length is what we need, then the
required angle, required pointing distance, and required thickness at
the base can be read off it.