Use absolute coordinates for find-intersections
Using "from" and "to" was a change, the previous code assumed the second
vector to be relative, so this breaks for other functions using it in
that way.

Also don't do the reverse sort of intersections found for a shape. This
currently won't work correctly and needs to be checked again, because
this way won't work for piles in sub-divisions when the origin is
outside of the shape.