Only consider the current environment for diagonal finding
Outside environments might be larger and shifted, resulting in the
middle of the line not being inside the desired environment.
This ensures the real start and end points are on the edge of the own
environment, which should be relatively safe.

One could still imagine a triangular or otherwise oddly shaped
environment, where the middle of the chosen start and end points ends up
outside of it, but for now this is the best idea I have.