Use bounding box instead of shape for per-bend intersections
This is much safer and easier, the bounding box intersections, which
must exist, are the best starting points for the line and much more
consistent to construct.

Additionally there've been some vector improvements. Intersections are
now pruned for duplicates, because the start/end point of a corner
yields two identical intersections.