Improve per-bend and per-bend-sinister further
Using the bounding box is a good start, it can help finding a middle
point that is expected to be inside the environment. From there use a
raw in both directions to find the first intersection with all
environments. This solves all sorts of problems with special cases.

Lines that start or end inside the environment will require a slight
adjustment to only shoot one ray.

Also fix the direction of the bends not being normalized, which resulted
in long lines, which were slowly rendered.