Refactor intersection logic again
This still isn't done, but I think it's shaping up now. Now
intersections are only considered edge intersections if they are inside
ALL environment shapes. Together with that and a check whether the
initial ray point is also inside all shapes or not, it is possible to
determine the first "outward" intersection. Namely, it'll be the first
if the point is inside, the second if the point is outside, because then
the first intersection is just the ray entering the shape.

This fixes the pile logic that the new intersection code broke.