Make bend and bend-sinister ordinaries work with full line styles
Also rewrite them entirely to not be rotated fess bands anymore. This is
necessary to find the intersections with the environment.
This also requires a change to the transform passed down the field, as
it must not transform the selection overlay.

Finally, the environment intersections now more correctly find all
intersections without the ray approach, the first and last then should
be the outer intersections the code needs.