Restructure dev/prod builds
This puts builds in a predictable place, more tidily separating dev and
prod builds.

Also separate dev and release build configuration in shadow-cljs.edn
directly, so it doesn't have to be done via --config-merge in the
Makefile.

Finally, the sharp copying stuff was moved into the deploy target, rather
than the shadow-cljs release build one, because it is only needed for
the serverless packaging.