Move :division into :field
Following the new spec from a previous commit, a field's :type now can
be :plain, then :tincture must be present, or the keyword for the
division, then :fields will contain the sub-fields, just like :division
:fields previously did.

All other options also moved from :division directly into the field map.

Some more polishing will be necessary to get the naming and namespace
separation right, now that "division" is not an entity in itself.
But these changes make the rendering and the forms functional with the
new spec.