Fix counterchanging with charge groups
Gotta clean up the context, as it is passed along and the charge-group
results in an endless loop when the charges copied inside the
counterchanged field are rendered.
The spacing passed along also confuses the automatic sizing of the
charges further down the rendering.

The :special point in the environment doesn't cause any direct problems,
but it's cleaner to not propagate it to other fields.