Display sub-fields and components as direct children of a field node
This reduces nesting a lot and actually makes a lot of sense. For the
current structure the user shouldn't need to care which group exactly an
element falls under. The buttons for adding a component can go into the
field node, where they also seem to make the most sense.