Simplify value-mode-select and other select fields
Each of them should get the options via subscription, reducing the
amount of arguments having to be passed all the way down the chain.

It's still useful to have separate element function that the multi
function calls, that way the element can be used directly, as in the
case of radio-select for charge-group-type-select, just overriding the
on-change handler.