Simplify UI ordinary-type-select, remove ordinary :type default
The value-mode-select is still added, but currently it shouldn't be
effective, as the ordinary's :type must not be nil and there currently is
no inheritance for ordinary types.

To make sure the value-mode-select element is ineffective, the :type
option for ordinaries must not have a default, which actually is correct in
general, as a lot of code is not prepared to handle a nil :type, since
it doesn't use option sanitization for it.

Once the data is properly sanitized, the default can be added and nil
values also will work in the coat of arms data for ordinaries, defaulting to
plain.