Change mirrored/flipped behaviour for secondary line styles
Previously these flags XORed the line's flags the secondary line derived
from, this allowed a default "false" to always have the main line's flag
set, and toggling it would invert the main line's flag, no matter what
it was.

This is magical stuff not covered by the options system, it requires
special logic in the rendering functions, and it's not obvious in the
interface. Also, it goes against the new inheritance system and
value-mode-select, which actually allows keeping the flag "nil" (Auto)
to match the main line's flag, this wasn't possible with a single,
boolean checkbox before.

So all that is thrown out and these flags can be inherited as other
values, and if main line and secondary line are supposed to have
different flags, then they need to be set manually.