Rewrite arms-library interface and CSS
The main view and the edit view have now been merged. This might not be
the final decision, but since I want to show the components it seems
silly to have just a single button switch between view/edit, while the
rest is the same.

The grid-based layout simplifies a lot of things, but the button row
still is a moving thing (depending on the size of the selected
component), and some elements have a border, others don't. This should
be unified somehow.

Finally, the attribution floats in the middle, without a separation line
of the grid area... perhaps such lines could structure the layout
better, but they also make it look boxed in and old-fashioned.