Improve :get-relevant-options again
The FIRST path going backwards that holds options information is the
deciding one. Currently there's no way anyway that an option higher up
could override values for this level, so the additional check of
relative-path being present is unnecessary.

As a side-effect the same path should always return the same
subscription for the first path going backwards that has option
information, so the problem the comment alludes to falls away.