Generalize render-options vs. render-options-path further
Now options/effective-value can be used for other things as well. The
big change is that only one value dictates how to proceed, if it is a
vector, then it is a path and :get-sanitized-value can be used,
otherwise it's assumed to be a map and needs to be sanitized first
before the value can be gotten.

This currently is less efficient, because in the latter case the
sanitizing of the entire root data is done again and again for each
attribute being processed.