WIP: start work on new approach that still allows re-frame + context input
The helper functions to access state and context likely won't stay like
this, but they'll allow the refactoring necessary.

tincture-select got a minor optimization, since it is special in not
changing any colours for the selected choice, the selection indicator
can be moved outside of the more expensive coat-of-arms rendering of the
choice, so the choice components never have to be re-rendered on a
change.