WIP: next steps in path-based-rendering rewrite
Abstract render-options into a special case of context-based
information, which reduces boiler plate code further and puts the logic
of how render-options are accessed and stored into fewer places.

This breaks most of the rendering code, so the plan is to add feature by
feature back with the new code.