Use better way to distinguish between option leaves and nodes
The value can be nil for both, so this is tricky. The idea here is that
option leaves have a :type key, but sometimes the option can be called
:type as well. So either there is no :type at all, then it is a node, or
an existing :type also has a :type key on its own, then it also is a
node, because this :type is not the option :type.