Add tinctured-field function that does path-based rendering
This allows changing the tincture without any re-rendering of parent
components.

In addition to that added a stable-id that uses sha1 with some stable
seeding input, which guarantees that the generated IDs will be stable
between multiple runs. That also will avoid some re-rendering of some
react components, such as tinctured-field, which takes an optional
mask-id.