Add mechanism to specify component-node paths that are open by default
The first 1-2 levels in the component trees currently in use should
always start open to avoid frustrating clicking. They still should be
collapsable, tho, to allow closing sub nodes easily.

That's why the node flag now also is stored as false, not removed,
allowing the distinction of "not yet touched" and "explicitly closed".