Make range input fields work with "incomplete" values
Now any typing is allowed, but the value is only parsed when the field
is blurred.

This allows typing "1." without the value being parsed back to "1" right
away, making it impossible to type a decimal number if the . isn't
already there.

This approach also allows updating the value when the hover-menu is used
to reset it, because tmp-value is only relevant while the field is
focussed.