Implement pin-reference? to fix collection element arms
This is incomplete, and I've decided not to implement this for now,
because it raises a bunch of other questions.

The current implementation will always store :version 0, so that's good,
and in most cases people will expect arms to update when they update
them, and without clear UI indication of what arms are being held back,
this would just be irritating.

I'll revert the commit afterwards, but this way it's still here when it
or parts of it become relevant again.