Rewrite achievement/coa/mottos resizing
This should be a more re-usable way to shift/scale a given bounding
box (coa + helms, then the resulting area of that + mottos) into a
target rectangle, given a target width and an optional constraint on the
aspect ratio.

Now the first resulting shift into a box already includes helms and
rotation, and the width of THAT rectangle can inform the mottos where to
go and what the target width is.