Rename vector +, -, *, / to add, sub, mul, div
The short function names were nice, but they required namespacing the
core +, -, *, / inside vector.core, and here and there errors happened
due to using the wrong one.

v+, etc. would be nice, but the symbol must not end in /, so that'd be a
special case. The three letter functions should be fine and always make
it clear which operation is done.