Refactor render passes to work with layer-separator
Now two pieces of information are passed down to the charge
rendering (the shared one for default charges and others): is it a
render pass below the shield? and is the charge itself below the shield?

Charges that don't have a layer-separator (and this is something
only the charge knows at render time right now) and that are not located
in the appropriate render pass (i.d. either it's below the shield and
the render pass isn't or vice versa) can render themselves as [:<>].

However, a charge that does have a layer-separator can render the parts
relevant to the render pass. Currently that means: below the shield
render everything but above the shield only render everything above the
layer separator.