Fix masking of overlay paths
This is glitchy due to a bug with how :clipPath works. Any pixel set, no
matter the colour is then considered as area to draw. The overlay
paths are usually not closed and they must not add to the mask, but even
a fill of "transparent" or "none" results in them being part of the
clipped area.

clipPath only was used to allow clicking through fields in the browser.
This feature currently is not used anyway, so use mask instead, which
does the right thing.

Do the same in heraldry.render, since it isn't used there either and
might prevent surprises.

This also results in a change in all generated option SVGs.