Use clipPath again for the main coat of arms
This reverts path of commit 3c7878393254ff824914af0ea6303ea59a608bc6.

Since the UI requires some click-detection on the coat of arms, most
notably in the collection library, here is a place where clipPath is
actually needed to avoid the annoying problems of a large, invisible
mask registering clicks in the wrong place.