Get a sanitized current value in ordinary options
Relying on a value or default might break if the given value is not
listed among available choices. So we need to check that as well.

This also avoids the awkward re-use of the default value. Cleaner and
easier to define the options once and use them to sanitize the value.