Combine interface/component-options and interface/options
They are similar and there was a lot of unnecessary dispatch logic.
Furthermore, the set of component type keyword is disjoint from the set
of entity type keywords, in fact, the keywords have different
namespaces, so the same function can be used.

To allow calling specific versions of the function a :dispatch-value can
be specified in the context, so different options can call other
versions, usually the version with entity type as dispatch value.