Combine all possible field option subscription into one set
This simplifies matters a bit and guarantees that the re-frame signal
graph inputs for the :options-for-component subscription will always be
the same for any path pointing to a field.

This is still fast, because the set is small-ish, however it means
sometimes options are computed again, even tho a change (e.g. the origin
point) won't change the options for that particular field. It's a
massive improvement nonetheless.