Make charge options subscription-based
Charge and ordinary need to have the same options because they can show
up on the same paths (i.e. in the components list of a field), so any
changes to that path must still have the right input subscriptions for
the re-frame signal graph.

Turn off charge validation for the time being. That'll get fixed once
everything is switched over.

Also remove anchor alignment for charge options, it doesn't make sense
there.