Move field inheritance into the rendering function
This allows honouring the flag (and the counterchanged type) of
referenced fields, it also simplifies environment management quite a
bit, as the override is not necessary.

This *might* be an issue again if environment calculations move into
subscriptions, but at that point this logic has to move there anyway, as
those subscriptions will have to follow the refs as well.