Improve :set-title for various entity forms
Turns out that the combination of :set-title setting ... :name and this
function subscribing to :name could create a loop. The loop didn't stop,
because the decision to :set-title was based on ... :id, not :name.

It's still not totally clear why this function/component re-renders on
some DB changes, even tho there are no more direct subscriptions in it,
but at least this stops the bleeding.