Shift submenu choice previews to align with the value
The translation values are a tad magical right now. They're simply the
rendered-choice-width * border-in-svg / viewport-width-in-svg in em.

This couples the choice images with the display here, but for now this
seems ok, as they don't often change anyway. At some point it might make
sense to store the choices without border and add the borders in the
submenu choice listing.