Refactor options-for-component once more
After the fix of getting around stale re-frame signal graph inputs, the
options for elements/components had to be recomputed on every db change
again, which is very inefficient.

This approach tries a middle way, in which option path and component
type are fetched outside the re-frame subscription, then passed along in
the context passed to the subscription. There they are used for the
options computation. This makes sure subscriptions update if any of
those values change, but as long as they stay the same, the re-frame
graph can work as intended.

This also means options and sanitized-data can't be used as input
subscriptions, unless the context has already been enriched with the
options-path and component type data. For that reason validations had to
be rewritten.

This increases performance a lot again, but it likely won't be the end
of the story. In particular it makes it hard to have layer 3
subscriptions that rely on that data, and a lot of UI parts could
benefit from them (validation, for one, submenu link names as well).

Finally, this new approach will allow splitting up
shared-options-subscriptions in a next step, which will significantly
reduce the signal graph complexity, as each component can narrow down
its own input subscriptions.